+++ aspnetcore/fundamentals/app-state/samples/3.x/SessionSample/Startup2.cs
@@ -48,13 +48,14 @@ namespace SessionSample
             app.UseHttpsRedirection();
             app.UseStaticFiles();
 
-            app.UseSession();
             app.UseHttpContextItemsMiddleware();
 
             app.UseRouting();
 
             app.UseAuthentication();
             app.UseAuthorization();
+
+            app.UseSession();
 
             app.UseEndpoints(endpoints =>
             {
