+++ aspnetcore/fundamentals/app-state/samples/3.x/SessionSample/Startup3.cs
@@ -40,18 +40,18 @@ namespace SessionSample
             app.UseHttpsRedirection();
             app.UseStaticFiles();
 
-            app.UseSession();
-
             app.UseRouting();
 
             app.UseAuthentication();
             app.UseAuthorization();
 
+            app.UseSession();
+
             app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapDefaultControllerRoute();
-                endpoints.MapRazorPages();
-            });
+    {
+        endpoints.MapDefaultControllerRoute();
+        endpoints.MapRazorPages();
+    });
         }
         #endregion
     }
